+++ packages/trueforge-sdk/src/serialization/types/SessionEvent.ts
@@ -14,6 +14,7 @@ import { ToolResponseEvent } from "./ToolResponseEvent.js";
 import { ToolResponseRequiredEvent } from "./ToolResponseRequiredEvent.js";
 import { TurnCreatedEvent } from "./TurnCreatedEvent.js";
 import { TurnDoneEvent } from "./TurnDoneEvent.js";
+import { TurnUpdateEvent } from "./TurnUpdateEvent.js";
 
 export const SessionEvent: core.serialization.Schema<serializers.SessionEvent.Raw, TrueForge.SessionEvent> =
     core.serialization.undiscriminatedUnion([
@@ -28,6 +29,7 @@ export const SessionEvent: core.serialization.Schema<serializers.SessionEvent.Ra
         ToolResponseRequiredEvent,
         TurnCreatedEvent,
         TurnDoneEvent,
+        TurnUpdateEvent,
     ]);
 
 export declare namespace SessionEvent {
@@ -42,5 +44,6 @@ export declare namespace SessionEvent {
         | ToolResponseEvent.Raw
         | ToolResponseRequiredEvent.Raw
         | TurnCreatedEvent.Raw
-        | TurnDoneEvent.Raw;
+        | TurnDoneEvent.Raw
+        | TurnUpdateEvent.Raw;
 }